CalculationController 이름 정정 및 RestController적용
